+++ Assets/Scripts/Components/Enemies/EnemyStats.cs
@@ -11,7 +11,7 @@ public class EnemyStats : InteractableStats
     [SerializeField] AudioClip deathSFX;
     [SerializeField] [Range(0, 1)] float deathSFXVolume = 0.7f;
 
-    public override void TakeDamage(int damage) {
+    public override void TakeDamage(float damage) {
         base.TakeDamage(damage);
         if (currentHealth <= 0) {
             Die();
